finished docs

+++ libafl/src/corpus/queue.rs
@@ -93,7 +93,7 @@ mod tests {
         let mut q =
             OnDiskCorpus::<BytesInput>::new(PathBuf::from("target/.test/fancy/path")).unwrap();
         let t = Testcase::with_filename(
-            BytesInput::new(vec![0 as u8; 4]),
+            BytesInput::new(vec![0_u8; 4]),
             "target/.test/fancy/path/fancyfile".into(),
         );
         q.add(t).unwrap();
